Upload artifacts directly from `acceptance` action (#69)
1. A new GitHub Action for Acceptance Testing has been added, which
executes tests, comments on PRs, links to workflow runs, and uploads
artifacts for later analysis. The action is designed to only create one
comment per PR, and to edit that comment with subsequent runs.
2. A new `project` input has been added to the `action.yml` file for the
acceptance test action, which allows users to specify the project name.
3. A new `uploadArtifact` function has been added to the `boilerplate`
package, which uploads artifacts to GitHub Actions. The function
prepares a temporary directory for artifact storage, creates a zip file
of the specified folder, uploads the zip file to GitHub Actions, and
returns metadata about the uploaded artifact.
4. A new `slog.go` file has been added to the `fixtures` package, which
initializes a logger for use in tests. The logger is configured to write
logs to a JSON file in the `DATABRICKS_LABS_LOG_DIR` directory.
5. The `init.go` file in the `fixtures` package has been updated to use
the new `slog.go` file to initialize the logger.
6. The `github` package has been updated with new methods for listing
commits and artifacts in a repository. The `ListCommits` method returns
a list of commits for a specified repository, branch, and path. The
`ListArtifacts` method returns a list of artifacts for a specified
repository.
7. The `github_actions.go` file in the `github` package has been updated
to include new fields in the `WorkflowRun` struct, such as `JobsURL`,
`LogsURL`, and `ArtifactsURL`.
8. The `sqlexec` package has been updated with new tests for error
mapping and multi-chunk operations. The `TestErrorMapping` test verifies
that the correct error is returned when a table already exists, and the
`TestMultiChunk` test verifies that the `InsertChunked` function can
insert data in multiple chunks.
9. The `comments.png` file is an example screenshot of a PR comment
created by the acceptance test action. The comment includes a summary of
the test results, a link to the workflow run, and a link to the uploaded
artifacts.
10. The `go.mod` and `go.sum` files have been updated to include new
dependencies. The `github.com/Azure/azure-sdk-for-go/sdk/storage/azblob`
package has been added to support the uploading of artifacts to GitHub
Actions. The `golang.org/x/net` and `golang.org/x/text` packages have
been added to support the `slog.go` file.